+++ DataManager/Models/UserModel.cs
@@ -18,7 +18,7 @@ namespace iRLeagueManager.Models
 
         private int? memberId;
         public int? MemberId { get => memberId; set => SetValue(ref memberId, value); }
-        int IAdmin.MemberId => (MemberId == null) ? 0 : (int)MemberId;
+        long IAdmin.MemberId => (MemberId == null) ? 0 : (int)MemberId;
 
         private string firstname;
         public string Firstname
